amend

+++ cleanrl/ppo_atari_envpool_torchcompile.py
@@ -249,7 +249,7 @@ if __name__ == "__main__":
 
     # TRY NOT TO MODIFY: start the game
     global_step = 0
-    next_obs = torch.tensor(envs.reset(), device=device)
+    next_obs = torch.tensor(envs.reset(), device=device, dtype=torch.float)
     next_done = torch.zeros(args.num_envs, device=device)
